+++ client/src/components/App.jsx
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from 'react';
-import Graph from './Graph.jsx';
+import GraphContainer from './GraphContainer.jsx';
 import WithLoading from './WithLoading.jsx';
 
 const App = () => {
-  const GraphWithLoading = WithLoading(Graph)
+  const GraphWithLoading = WithLoading(GraphContainer)
   const [appState, setAppState] = useState({
     loading: true
   });
