+++ package.json
@@ -17,11 +17,13 @@
     "@babel/preset-env": "^7.17.12",
     "@babel/preset-react": "^7.16.5",
     "@babel/runtime": "^7.17.9",
+    "@svgr/webpack": "^6.2.1",
     "babel-loader": "^8.2.3",
     "css-loader": "^6.7.1",
     "eslint": "^8.15.0",
     "eslint-plugin-editorconfig": "^3.2.0",
     "style-loader": "^3.3.1",
+    "svg-inline-loader": "^0.8.2",
     "webpack": "^5.65.0",
     "webpack-cli": "^4.9.1"
   },
